添加 JobsMW 测试
