+++ jest.config.js
@@ -7,24 +7,25 @@ if (fs.existsSync('.env')) {
 }
 
 module.exports = {
-    moduleNameMapper: {
-        '^@/(.*)$': '<rootDir>/$1'
-    },
-    moduleFileExtensions: [
-        'ts',
-        'js',
-        'json'
-    ],
-    transform: {
-        "^.+\\.ts$": "ts-jest",
-        '^.+\\.js$': 'babel-jest',
-    },
-    collectCoverage: true,
-    collectCoverageFrom: [
-      '<rootDir>/scripts/**/*.ts',
-      '<rootDir>/controllers/**/*.ts',
-        '<rootDir>/routes/**/*.ts',
-        "!**/node_modules/**"
-    ],
-    coverageReporters: ["html", "text-summary", "lcov"],
+  moduleNameMapper: {
+    '^@/(.*)$': '<rootDir>/$1'
+  },
+  moduleFileExtensions: [
+    'ts',
+    'js',
+    'json'
+  ],
+  transform: {
+    "^.+\\.ts$": "ts-jest",
+    '^.+\\.js$': 'babel-jest',
+  },
+  collectCoverage: true,
+  collectCoverageFrom: [
+    '<rootDir>/scripts/**/*.ts',
+    '<rootDir>/middlewares/**/*.ts',
+    '<rootDir>/controllers/**/*.ts',
+    '<rootDir>/routes/**/*.ts',
+    "!**/node_modules/**"
+  ],
+  coverageReporters: ["html", "text-summary", "lcov"],
 }
